app/src/test/java/com/onoh/moviecataloeguerepositoryarch/ui/dashboard/movie/MovieViewModelTest.kt
package com.onoh.moviecataloeguerepositoryarch.ui.dashboard.movie

import androidx.arch.core.executor.testing.InstantTaskExecutorRule
import androidx.lifecycle.MutableLiveData
import androidx.lifecycle.Observer
import androidx.paging.PagedList
import com.onoh.moviecataloeguerepositoryarch.data.AppRepository
import com.onoh.moviecataloeguerepositoryarch.data.local.entity.MovieEntity
import com.onoh.moviecataloeguerepositoryarch.vo.Resource
import org.junit.Assert
import org.junit.Before
import org.junit.Rule
import org.junit.Test
import org.junit.runner.RunWith
import org.mockito.Mock
import org.mockito.Mockito
import org.mockito.junit.MockitoJUnitRunner

@RunWith(MockitoJUnitRunner::class)
class MovieViewModelTest {
//    -obj
//    1.Data movie tidak null
//    2.Memastikan jumlah data movie sesuai dengan yang diharapkan

    //    objective tes ke 2
//    Memanipulasi data ketika pemanggilan data movie di kelas repository.
//    Memastikan metode di kelas repository terpanggil.
//    Melakukan pengecekan data movie apakah null atau tidak.
//    Melakukan pengecekan jumlah data movie apakah sudah sesuai atau belum

    private lateinit var viewModel: MovieViewModel

    @get:Rule
    var instantTaskExecutorRule = InstantTaskExecutorRule()

    @Mock
    private lateinit var appRepository: AppRepository

    @Mock
    private lateinit var observer: Observer<Resource<PagedList<MovieEntity>>>

    @Mock
    private lateinit var pagedList: PagedList<MovieEntity>

    @Before
    fun setUp() {
        viewModel = MovieViewModel(appRepository)
    }

    @Test
    fun getMovies() {
        val dummyMovie = Resource.success(pagedList)
        Mockito.`when`(dummyMovie.data?.size).thenReturn(20)
        val movies = MutableLiveData<Resource<PagedList<MovieEntity>>>()
        movies.value = dummyMovie

        Mockito.`when`(appRepository.getAllMovies()).thenReturn(movies)
        val movieEntities = viewModel.getMovies().value?.data
        Mockito.verify(appRepository).getAllMovies()
        Assert.assertNotNull(movieEntities)
        Assert.assertEquals(20, movieEntities?.size)

        viewModel.getMovies().observeForever(observer)
        Mockito.verify(observer).onChanged(dummyMovie)


    }
}